Isolate Payment Fixed Battlesim. Fuxed unlinks errors

+++ trunk/includes/classes/class.statbuilder.php
@@ -88,7 +88,8 @@ class statbuilder extends records
 		{
 			while($RID = $db->fetch_array($DelRW))
 			{
-				unlink(ROOT_PATH.'raports/raport_'.$RID['rid'].'.php');
+				if(file_exits(ROOT_PATH.'raports/topkb_'.$RID['rid'].'.php'))
+					unlink(ROOT_PATH.'raports/raport_'.$RID['rid'].'.php');
 			}	
 			$db->query("DELETE FROM ".RW." WHERE `time` < '". $del_before ."';");
 		}
@@ -102,7 +103,8 @@ class statbuilder extends records
 			{
 				while($RID = $db->fetch_array($TKBRW))
 				{
-					unlink(ROOT_PATH.'raports/topkb_'.$RID['rid'].'.php');
+					if(file_exits(ROOT_PATH.'raports/topkb_'.$RID['rid'].'.php'))
+						unlink(ROOT_PATH.'raports/topkb_'.$RID['rid'].'.php');
 				}	
 				$db->query("DELETE FROM ".TOPKB." WHERE `gesamtunits` < '".((isset($TopKBLow)) ? $TopKBLow['gesamtunits'] : 0)."';");
 			}
